scraper/requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.11.1
pandas==2.0.3
pydantic==2.0.2
Requests==2.31.0
SQLAlchemy==2.0.18
psycopg2-binary==2.9.6

